oasp/oasp4j#304: NPE fixes + retrieve raw picture service added

+++ oasp4j-samples/oasp4j-sample-core/src/main/java/io/oasp/gastronomy/restaurant/offermanagement/logic/impl/OffermanagementImpl.java
@@ -6,6 +6,7 @@ import io.oasp.gastronomy.restaurant.general.logic.base.AbstractComponentFacade;
 import io.oasp.gastronomy.restaurant.general.logic.base.UcManageBinaryObject;
 import io.oasp.gastronomy.restaurant.offermanagement.common.api.datatype.ProductType;
 import io.oasp.gastronomy.restaurant.offermanagement.common.api.exception.OfferEmptyException;
+import io.oasp.gastronomy.restaurant.offermanagement.common.api.exception.ProductNotFoundException;
 import io.oasp.gastronomy.restaurant.offermanagement.dataaccess.api.OfferEntity;
 import io.oasp.gastronomy.restaurant.offermanagement.dataaccess.api.ProductEntity;
 import io.oasp.gastronomy.restaurant.offermanagement.dataaccess.api.dao.DrinkDao;
@@ -321,14 +322,24 @@ public class OffermanagementImpl extends AbstractComponentFacade implements Offe
   @RolesAllowed(PermissionConstants.FIND_PRODUCT)
   public BinaryObjectEto findProductPicture(Long productId) {
 
-    return getUcManageBinaryObject().findBinaryObject(findProduct(productId).getPictureId());
+    ProductEto product = findProduct(productId);
+    if (product != null) {
+      return getUcManageBinaryObject().findBinaryObject(product.getPictureId());
+    } else {
+      return null;
+    }
   }
 
   @Override
   @RolesAllowed(PermissionConstants.FIND_PRODUCT_PICTURE)
   public Blob findProductPictureBlob(Long productId) {
 
-    return getUcManageBinaryObject().getBinaryObjectBlob(findProductPicture(productId).getId());
+    ProductEto product = findProduct(productId);
+    if (product != null) {
+      return getUcManageBinaryObject().getBinaryObjectBlob(product.getPictureId());
+    } else {
+      return null;
+    }
   }
 
   @Override
@@ -366,9 +377,13 @@ public class OffermanagementImpl extends AbstractComponentFacade implements Offe
   public void updateProductPicture(Long productId, Blob blob, BinaryObjectEto binaryObjectEto) {
 
     ProductEntity product = getProductDao().findOne(productId);
-    binaryObjectEto = getUcManageBinaryObject().saveBinaryObject(blob, binaryObjectEto);
-    product.setPictureId(binaryObjectEto.getId());
-    getProductDao().save(product);
+    if (product != null) {
+      binaryObjectEto = getUcManageBinaryObject().saveBinaryObject(blob, binaryObjectEto);
+      product.setPictureId(binaryObjectEto.getId());
+      getProductDao().save(product);
+    } else {
+      throw new ProductNotFoundException(productId);
+    }
 
   }
 
@@ -377,8 +392,9 @@ public class OffermanagementImpl extends AbstractComponentFacade implements Offe
   public void deleteProductPicture(Long productId) {
 
     ProductEntity product = getProductDao().findOne(productId);
-    getUcManageBinaryObject().deleteBinaryObject(product.getPictureId());
-
+    if (product != null) {
+      getUcManageBinaryObject().deleteBinaryObject(product.getPictureId());
+    }
   }
 
   @Override
